create method ok

+++ src/main/java/com/example/Aerolinea/repositories/IDestinationRepository.java
@@ -3,5 +3,5 @@ package com.example.Aerolinea.repositories;
 import com.example.Aerolinea.model.Destination;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IDestinationRepository extends JpaRepository<Destination, Integer>{
+public interface IDestinationRepository extends JpaRepository<Destination, Long>{
 }
